docs: Change code snippets to appear inline
Slate places code snippets in the right half of the page, which
is dedicated just to code snippets. This doesn't make sense for
our docs, for which it's easier to read the code snippets inline.
This commit removes support for the dedicated code area on the
right hand side, and adds some formatting so that the inline
code blocks look reasonable.